requirements.txt
mcp-use @ git+https://github.com/mcp-use/mcp-use.git@fix/transport-security-host-passthrough-v2#subdirectory=libraries/python
mcp>=1.0.0
google-genai>=1.7.0
pillow>=11.1.0
python-dotenv>=1.0.0
